+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-build: Fixed showing the wrong indices in weapon_hud_interface errors
 - invader-build: Fixed non-power-of-two check giving incorrect results when
   compiling bitmaps
+- invader-build: Fixed a missing "bip01 head" resulting in a warning
 
 ## [0.25.0] - 2020-02-11
 ### Added